src/third_party/blink/web_tests/external/wpt/css/css-transforms/transforms-support-calc-expected.txt

This is a testharness.js-based test.
FAIL translate supports calc assert_equals: expected "calc(20% + 30px) calc(100% - 200px)" but got "calc(20% + 30px) calc(100% + -200px)"
PASS rotate supports calc
PASS scale supports calc
PASS perspective supports calc
PASS perspective-origin supports calc
PASS transform supports calc
PASS transform-origin supports calc
Harness: the test ran to completion.
